+++ gas-measurement.sh
@@ -6,7 +6,7 @@ default_json_file="${out_dir}/gas-measurement.json"
 function run_gas_measurement() {
   local json_file=$1
   local tmp_file="$1.tmp"
-  forge test --mp "test/core/gas/*" -vv --silent --json > "${json_file}"
+  forge test --mp "$GAS_MATCH_PATH" -vv --silent --json > "${json_file}"
   # Forge is very verbose, so we filter the json to keep only the gas information
   jq '
     to_entries |